final edit isa
